+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 * jdbc: let tryLock/unlock not eat exceptions; also log if unlock returns false
+* jdbc: insertBatch()/upsertBatch() now work properly with mixing of nullable and not nullable @JsonColumn
 * jobs: fix locking/unlocking of non-transactional jobs that should not run in parallel
 * json: flush partial buffer when generating of json fails
 * openapi: guarantee that /openapi.json is rendered as json, no matter which renderers are used in the project